Refactor: 초기 코드 정리
- 서비스 메소드에 주석 추가
- 불필요한 파일 정리
- 요청/응답 데이터 패키지 분리
- 예외 패키지명, 어드바이스 클래스명 수정
