+++ app/src/main/java/com/codesoom/assignment/application/UserService.java
@@ -2,10 +2,10 @@ package com.codesoom.assignment.application;
 
 import com.codesoom.assignment.domain.User;
 import com.codesoom.assignment.domain.UserRepository;
-import com.codesoom.assignment.dto.UserModificationData;
-import com.codesoom.assignment.dto.UserRegistrationData;
-import com.codesoom.assignment.errors.UserEmailDuplicationException;
-import com.codesoom.assignment.errors.UserNotFoundException;
+import com.codesoom.assignment.dto.request.UserModificationData;
+import com.codesoom.assignment.dto.request.UserRegistrationData;
+import com.codesoom.assignment.exception.UserEmailDuplicationException;
+import com.codesoom.assignment.exception.UserNotFoundException;
 import com.github.dozermapper.core.Mapper;
 import org.springframework.stereotype.Service;
 
@@ -22,6 +22,12 @@ public class UserService {
         this.userRepository = userRepository;
     }
 
+    /**
+     * 회원을 저장합니다.
+     * @param registrationData 저장할 회원
+     * @return 저장된 회원
+     * @throws UserEmailDuplicationException 이메일이 기존에 존재하는 경우
+     */
     public User registerUser(UserRegistrationData registrationData) {
         String email = registrationData.getEmail();
         if (userRepository.existsByEmail(email)) {
@@ -32,6 +38,13 @@ public class UserService {
         return userRepository.save(user);
     }
 
+    /**
+     * 회원을 수정합니다.
+     * @param id 수정할 회원 아이디
+     * @param modificationData 수정할 회원
+     * @return 수정된 회원
+     * @throws UserNotFoundException 수정할 회원을 찾지 못한 경우
+     */
     public User updateUser(Long id, UserModificationData modificationData) {
         User user = findUser(id);
 
@@ -41,12 +54,24 @@ public class UserService {
         return user;
     }
 
+    /**
+     * 회원을 삭제합니다.
+     * @param id 삭제할 회원 아이디
+     * @return 삭제된 회원
+     * @throws UserNotFoundException 삭제할 회원을 찾지 못한 경우
+     */
     public User deleteUser(Long id) {
         User user = findUser(id);
         user.destroy();
         return user;
     }
 
+    /**
+     * 회원을 조회하여 반환합니다.
+     * @param id 조회할 회원 아이디
+     * @return 조회된 회원
+     * @throws UserNotFoundException 회원을 찾지 못한 경우
+     */
     private User findUser(Long id) {
         return userRepository.findByIdAndDeletedIsFalse(id)
                 .orElseThrow(() -> new UserNotFoundException(id));
