--- app/src/main/java/com/codesoom/assignment/controllers/HelloController.java
@@ -1,12 +1,0 @@
-package com.codesoom.assignment.controllers;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class HelloController {
-    @RequestMapping("/")
-    public String sayHello() {
-        return "Hello, world!";
-    }
-}
