+++ app/src/main/java/com/codesoom/assignment/controllers/UserController.java
@@ -2,9 +2,9 @@ package com.codesoom.assignment.controllers;
 
 import com.codesoom.assignment.application.UserService;
 import com.codesoom.assignment.domain.User;
-import com.codesoom.assignment.dto.UserModificationData;
-import com.codesoom.assignment.dto.UserRegistrationData;
-import com.codesoom.assignment.dto.UserResultData;
+import com.codesoom.assignment.dto.request.UserModificationData;
+import com.codesoom.assignment.dto.request.UserRegistrationData;
+import com.codesoom.assignment.dto.response.UserResultData;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
