+++ app/src/main/java/com/codesoom/assignment/domain/ProductRepository.java
@@ -1,9 +1,11 @@
 package com.codesoom.assignment.domain;
 
+import org.springframework.data.jpa.repository.JpaRepository;
+
 import java.util.List;
 import java.util.Optional;
 
-public interface ProductRepository {
+public interface ProductRepository extends JpaRepository<Product, Long> {
     List<Product> findAll();
 
     Optional<Product> findById(Long id);
