+++ app/src/main/java/com/codesoom/assignment/domain/UserRepository.java
@@ -1,8 +1,10 @@
 package com.codesoom.assignment.domain;
 
+import org.springframework.data.jpa.repository.JpaRepository;
+
 import java.util.Optional;
 
-public interface UserRepository {
+public interface UserRepository extends JpaRepository<User, Long> {
     User save(User user);
 
     boolean existsByEmail(String email);
