+++ app/src/main/java/com/codesoom/assignment/dto/request/UserModificationData.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.dto;
+package com.codesoom.assignment.dto.request;
 
 import com.github.dozermapper.core.Mapping;
 import lombok.AllArgsConstructor;
