--- app/src/main/java/com/codesoom/assignment/infra/JpaUserRepository.java
@@ -1,20 +1,0 @@
-package com.codesoom.assignment.infra;
-
-import com.codesoom.assignment.domain.User;
-import com.codesoom.assignment.domain.UserRepository;
-import org.springframework.data.repository.CrudRepository;
-
-import java.util.Optional;
-
-public interface JpaUserRepository
-        extends UserRepository, CrudRepository<User, Long> {
-    User save(User user);
-
-    boolean existsByEmail(String email);
-
-    Optional<User> findById(Long id);
-
-    Optional<User> findByIdAndDeletedIsFalse(Long id);
-
-    Optional<User> findByEmail(String email);
-}
